♻️ WSRP8V pr: decompose flow status rendering
