+++ .agentplane/tasks/202605290543-WSRP8V/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 WSRP8V task: PR flow status render decomposition [202605290543-WSRP8V]
