+++ .agentplane/tasks/202605290543-WSRP8V/pr/review.md
@@ -1,0 +1,38 @@
+# PR Review
+
+Created: 2026-05-29T05:43:31.527Z
+
+## Task
+
+- Task: `202605290543-WSRP8V`
+- Title: PR flow status render decomposition
+- Status: DOING
+- Branch: `task/202605290543-WSRP8V/pr-flow-status-render-decomposition`
+- Canonical task record: `.agentplane/tasks/202605290543-WSRP8V/README.md`
+
+## Verification
+
+- State: ok
+- Note: PR flow status text rendering extracted into flow-status.render.ts; flow-status.ts reduced to 367 lines while preserving report resolution and output rows.
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T05:43:31.527Z
+- Branch: task/202605290543-WSRP8V/pr-flow-status-render-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+ .../src/commands/pr/flow-status.render.ts          | 66 ++++++++++++++++++++++
+ packages/agentplane/src/commands/pr/flow-status.ts | 59 +------------------
+ 2 files changed, 68 insertions(+), 57 deletions(-)
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
